content/media/mediasource/test/test_SeekableBeforeEndOfStream.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seekable attribute before end of stream</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(function (ms, v) {
  ms.addEventListener("sourceopen", function () {
    var sb = ms.addSourceBuffer("video/webm");

    fetchWithXHR("seek.webm", function (arrayBuffer) {
      sb.appendBuffer(new Uint8Array(arrayBuffer));
    });

    var target = 2;

    v.addEventListener("loadedmetadata", function () {
      ok(v.seekable.length, "Resource is seekable");
      ok(v.seekable.length &&
          target >= v.seekable.start(0) &&
          target < v.seekable.end(0), "Target is within seekable range");
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
